Convert ChConstraintThree and derived class to double precision

+++ src/chrono/solver/ChConstraintThree.h
@@ -53,18 +53,18 @@ class ChApi ChConstraintThree : public ChConstraint {
     ChConstraintThree& operator=(const ChConstraintThree& other);
 
     /// Access jacobian matrix
-    virtual ChMatrix<float>* Get_Cq_a() = 0;
+    virtual ChMatrix<double>* Get_Cq_a() = 0;
     /// Access jacobian matrix
-    virtual ChMatrix<float>* Get_Cq_b() = 0;
+    virtual ChMatrix<double>* Get_Cq_b() = 0;
     /// Access jacobian matrix
-    virtual ChMatrix<float>* Get_Cq_c() = 0;
+    virtual ChMatrix<double>* Get_Cq_c() = 0;
 
     /// Access auxiliary matrix (ex: used by iterative solvers)
-    virtual ChMatrix<float>* Get_Eq_a() = 0;
+    virtual ChMatrix<double>* Get_Eq_a() = 0;
     /// Access auxiliary matrix (ex: used by iterative solvers)
-    virtual ChMatrix<float>* Get_Eq_b() = 0;
+    virtual ChMatrix<double>* Get_Eq_b() = 0;
     /// Access auxiliary matrix (ex: used by iterative solvers)
-    virtual ChMatrix<float>* Get_Eq_c() = 0;
+    virtual ChMatrix<double>* Get_Eq_c() = 0;
 
     /// Access the first variable object
     ChVariables* GetVariables_a() { return variables_a; }
